+++ src/pages/Dashboard/SideBar/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { FiBarChart2 } from 'react-icons/fi';
 
 import { Container, AmountInvested } from './styles';
-import { dashboardNavigation } from '../../../resources';
+import { sideBarNavigation } from '../../../resources';
 import Dropdown from './Dropdown';
 import AccountBalance from './AccountBalance';
 
@@ -14,7 +14,7 @@ const SideBar: React.FC = () => {
       <AmountInvested>
         Ver Total Investido <FiBarChart2 size="1.6rem" />
       </AmountInvested>
-      {Object.entries(dashboardNavigation).map(([key, value]) => (
+      {Object.entries(sideBarNavigation).map(([key, value]) => (
         <Dropdown key={key} icon={value.icon} sectionName={key} />
       ))}
     </Container>
